internal/nlp/requirements.txt
annotated-types==0.7.0
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
grpcio==1.71.0
grpcio-tools==1.71.0
idna==3.10
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
numpy==2.0.2
packaging==24.2
preshed==3.0.9
protobuf==5.29.4
pydantic==2.11.0
pydantic_core==2.33.0
Pygments==2.19.1
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
smart-open==7.1.0
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.3.4
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
